pyproject.toml
[project]
name = "networker"
version = "0.1.0"
description = "Telegram bot for networking at hackathons"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk==3.5.0",
    "google-auth>=2.40.3",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
]
